+++ internal/cmd/firmware/get.go
@@ -1,0 +1,37 @@
+package firmware
+
+import (
+	"fmt"
+	"net/url"
+
+	"github.com/spf13/cobra"
+
+	"github.com/inhandnet/devicemanager-cli/internal/factory"
+	"github.com/inhandnet/devicemanager-cli/internal/iostreams"
+)
+
+func NewCmdGet(f *factory.Factory) *cobra.Command {
+	return &cobra.Command{
+		Use:   "get <firmware-id>",
+		Short: "Get firmware details",
+		Args:  cobra.ExactArgs(1),
+		RunE: func(cmd *cobra.Command, args []string) error {
+			client, err := f.APIClient()
+			if err != nil {
+				return err
+			}
+
+			q := url.Values{}
+			q.Set("verbose", "100")
+
+			output, _ := cmd.Flags().GetString("output")
+
+			body, err := client.Get(fmt.Sprintf("/api/firmware/%s", args[0]), q)
+			if err != nil {
+				return err
+			}
+
+			return iostreams.FormatOutput(body, f.IO, output)
+		},
+	}
+}