+++ internal/cmd/system/role.go
@@ -1,0 +1,48 @@
+package system
+
+import (
+	"net/url"
+
+	"github.com/spf13/cobra"
+
+	"github.com/inhandnet/devicemanager-cli/internal/factory"
+	"github.com/inhandnet/devicemanager-cli/internal/iostreams"
+)
+
+func NewCmdRole(f *factory.Factory) *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "role",
+		Short: "Manage roles",
+	}
+
+	cmd.AddCommand(newCmdRoleList(f))
+
+	return cmd
+}
+
+func newCmdRoleList(f *factory.Factory) *cobra.Command {
+	return &cobra.Command{
+		Use:     "list",
+		Short:   "List roles in the organization",
+		Aliases: []string{"ls"},
+		RunE: func(cmd *cobra.Command, args []string) error {
+			client, err := f.APIClient()
+			if err != nil {
+				return err
+			}
+
+			q := url.Values{}
+			q.Set("verbose", "100")
+
+			output, _ := cmd.Flags().GetString("output")
+
+			body, err := client.Get("/api2/roles", q)
+			if err != nil {
+				return err
+			}
+
+			return iostreams.FormatOutput(body, f.IO, output,
+				iostreams.WithColumns("_id", "name", "type"))
+		},
+	}
+}